requirements.txt
flask
transformers
torch
datasets
pandas
scikit-learn
requests
beautifulsoup4
python-dotenv
Pillow
gunicorn